refactoring to reuse the core data persistence container
